requirements.txt
torch
faster-whisper
fastapi
uvicorn
websockets
python-multipart
numpy
pyannote.audio
python-dotenv